n_avanzado/requirements.txt
streamlit
google-cloud-aiplatform
google-cloud
python-dotenv
requests
vertexai
langchain
langchain-community
transformers
diffusers
torch
accelerate
sentencepiece
sentence-transformers
pillow
arxiv
faiss-cpu